# Back-end/server/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "UsersController_getAllUsers",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "user role (citizen, officer, supervisor, superuser)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "user role (citizen, officer, supervisor, superuser)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "user role (citizen, officer, supervisor, superuser)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "user role (citizen, officer, supervisor, superuser)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "user role (citizen, officer, supervisor, superuser)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/register": {
      "post": {
        "operationId": "UsersController_createCitizen",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "user role (citizen, officer, supervisor, superuser)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register as citizen",
        "tags": [
          "users"
        ]
      }
    },
    "/cases": {
      "post": {
        "operationId": "CasesController_createCase",
        "parameters": [
          {
            "name": "userid",
            "in": "header",
            "description": "user id for filtering (citizen/officer)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "citizen | officer | supervisor | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Citizen creates a case",
        "tags": [
          "cases"
        ]
      },
      "get": {
        "operationId": "CasesController_getAllCases",
        "parameters": [
          {
            "name": "userid",
            "in": "header",
            "description": "user id for filtering (citizen/officer)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "citizen | officer | supervisor | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userid",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get cases (role-based)",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{id}": {
      "get": {
        "operationId": "CasesController_getCase",
        "parameters": [
          {
            "name": "userid",
            "in": "header",
            "description": "user id for filtering (citizen/officer)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "citizen | officer | supervisor | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get case by ID",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{id}/status": {
      "patch": {
        "operationId": "CasesController_updateStatus",
        "parameters": [
          {
            "name": "userid",
            "in": "header",
            "description": "user id for filtering (citizen/officer)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "citizen | officer | supervisor | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Officer updates case status",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{id}/assign": {
      "patch": {
        "operationId": "CasesController_assignCase",
        "parameters": [
          {
            "name": "userid",
            "in": "header",
            "description": "user id for filtering (citizen/officer)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "citizen | officer | supervisor | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supervisor assigns case",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{id}/request-closure": {
      "patch": {
        "operationId": "CasesController_requestClosure",
        "parameters": [
          {
            "name": "userid",
            "in": "header",
            "description": "user id for filtering (citizen/officer)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "citizen | officer | supervisor | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Officer requests closure",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{id}/closure-decision": {
      "patch": {
        "operationId": "CasesController_handleClosure",
        "parameters": [
          {
            "name": "userid",
            "in": "header",
            "description": "user id for filtering (citizen/officer)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "citizen | officer | supervisor | superuser",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClosureDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Supervisor handles closure request",
        "tags": [
          "cases"
        ]
      }
    }
  },
  "info": {
    "title": "Case Management API",
    "description": "Case Management API description",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCaseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateStatusDto": {
        "type": "object",
        "properties": {}
      },
      "AssignCaseDto": {
        "type": "object",
        "properties": {}
      },
      "ClosureDecisionDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}